Add the "Active users" meter.
A text meter that display a list of users owning processes that are actually running.
The list is truncated to the size of the display.
Color active users according to levels of activity.

Users are categorized in 3 classes: high, middle and low activity.
Users are displayed in descending order of activity, with different colors.
Users shown in higher levels are not shown in lower ones, even if they have active processes in those levels.
Bounds of the level classes are given as percentages: >95%, 95-50%, <50%.

Data are stored in a NEW Meter.strings array, that behaves just like the Meter.values one,
but that handle strings instead of doubles.
Remove the static attribute for RichString_writeFrom, because it is used by this module.
Adds ACTIVE_USERS_* style fields.