Color active users according to levels of activity.
This is an update of the text meter that shows the names of users having active processes.
Users are categorized in 3 classes: high, middle and low activity.
Users are displayed in descending order of activity, with different colors.
Users shown in higher levels are not shown in lower ones, even if they have active processes in those levels.
Bounds of the level classes are given as percentages: >95%, 95-50%, <50%.

Data are stored in a NEW Meter.strings array, that behaves just like the Meter.values one,
but that handle strings instead of doubles.

Remove the static attribute for RichString_writeFrom, because it is used by this module.
Adds ACTIVE_USERS_* style fields.