[Experimental] Graph meter coloring by items.
This patch colors the graph meter based on the items of the bar meter.
For example, if most of the CPU power is running on low-priority
processes, then blue will be shown on the graph columns. If most of the
CPU is running on kernel things, then the columns will be red, and so
on.

The algorithm in this patch is not final yet. Expect some bugs such as:
that some of the cells are colored gray while ideally they shouldn't.