Graph meter coloring by items. (Rebased commit, part 1)
This patch colors the graph meter based on the items of the bar meter.
For example, if most of the CPU power is running on low-priority
processes, then blue will be shown on the graph columns. If most of CPU
is running on kernel things, then the columns will be red, and so on.

This commit contains a simple, yet quadratic time algorithm for
determining color.

Additional changes:
* Adjusts printing of graph bars from top down to bottom up.
* Minor formatting changes in Meter.c.