RFC: silence compiler warnings
For more recent GCC versions casting to (void) is not sufficient to
silence the warning about ignoring return values.

In the past we assigned the return value to a variable and casted that
to (void). Not really an elegant method.

How about mis-using a condition with empty body? This is just a RFC,
if you are happy I will catch some more cases and update the patch.