Turn Vector::swap() into a static inline function.
Should result in the same binary. If not, it's a compiler bug.

As a macro, its declaration and usage differ from convention. This
swap() allows swapping values inside the same array only, while
conventional swap macro allows swapping any pair of variables of the
same type. Therefore, it's better not expose it in the header file,
lest it be used by other htop code.

Signed-off-by: Kang-Che Sung <explorer09 @ gmail.com>