Limit signal list display to first 99 signals.
This reverts "Protect against strange values of SIGRTMIN and SIGRTMAX,
and change alignment of numbers." (5df425867cd992ccadbe8b5cf713404fd53db7df)
and changes it to a (I believed) better bound protection. Now the list
only displays first 99 signals that are available.

In case of a system providing so called "strange" or too large SIGRTMAX
value, let it be. That may not be an error but an OS feature.

Also note the "%+d" to "%-+3d" format change was technically redundant.