Rename Meter.total variable to Meter.full
The word "total" has a dictionary meaning of "sum" or "a quantity
obtained by addition of a group of members" [WordNet]. Although this
term works when we talk about "total CPU time" or "total memory", it
may be inadequate to the context of meter drawing. For example,
TasksMeter fills this variable with total number of processes; while
this fits the dictionary meaning of "total", it will go weird when it
is plotted as a graph. (And it will confuse my graph coloring
algorithm, too.)

Rename this variable to "full" to better reflect it's use - indicating
the "full" value of the bar/graph.

Signed-off-by: Kang-Che Sung <explorer09@gmail.com>