Generalize timeradd change
timeradd is a macro everywhere I've seen it, so it should be fine.
Maybe not as scaleable if more usage of timeradd is employed, but
oh well.

I tried adding the OpenBSD definition for timeradd and using ifdefs
for it, but bizarrely, even after defining timeradd, it refused to
use the definition, instead trying an implicit function declaration
instead. (???) This should work in the meantime.