Change OpenBSD thread display, including hackish workaround for hash table lookups returning strange results.
Modify the thread display so that it does not display child threads in
the case where a process is single-threaded. This is achieved by first
looking up every process in the system, and then looking up how many
threads each process has.

Something somewhere is subtly broken in ways which I don't understand.
On 32-bit PowerPC (big-endian), this code runs without any issues (after
applying the OpenBSD ports patches), however on 64-bit Intel
(little-endian) htop will frequently crash when attempting to
dereference a pointer with the value 0x1 in
OpenBSDProcessList_goThroughEntries, which is returned from the hash
table lookup in ProcessList_getProcess. This appears to be related to
the process hash table load factor, as increasing the number of buckets
reduces the frequency of crashes, and reducing the number of buckets
will eventually cause htop to reliably crash on the first screen
refresh.

I did occasionally observe this problem on the upstream master branch,
but I had difficulty reproducing it (though this is probably because
counting all the threads on the system increases the number of entries
in the hash table enough to trigger this issue). I have added a check to
OpenBSDProcessList_goThroughEntries to skip to the next process if
ProcessList_getProcess returns (Process *) 0x1, but this is a massive
hack.